feat(products): add products api view

+++ food_project/products/views.py
@@ -1,4 +1,4 @@
-from rest_framework.views import APIView
+from rest_framework.decorators import api_view
 from rest_framework.generics import get_object_or_404
 from rest_framework.response import Response
 from rest_framework import status
@@ -7,26 +7,21 @@ from .serializers import ProductSerializer
 
 
 # Create your views here.
-class ProductAPIView(APIView):
+@api_view(["GET", "PUT", "DELETE"])
+def product_api_view(request, pk):
+    product = Product.objects.get(pk=pk)
 
-    def get_object(self, pk):
-        product = get_object_or_404(Product, pk=pk)
-        return product
-
-    def get(self, request, pk):
-        product = self.get_object(pk)
-        serializer = ProductSerializer(peoduct)
+    if request.method == "GET":
+        serializer = ProductSerializer(product)
         return Response(serializer.data)
 
-    def put(self, request, pk):
-        product = self.get_object(pk)
+    elif request.method == "PUT":
         serializer = ProductSerializer(product, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, request, pk):
-        product = self.get_object(pk)
+    elif request.method == "DELETE":
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
